reorganization of classes into better folder structure
